+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/BlobServiceClientBuilder.java
@@ -13,8 +13,11 @@ import com.azure.core.util.CoreUtils;
 import com.azure.core.util.Configuration;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.storage.blob.implementation.util.BuilderHelper;
+import com.azure.storage.blob.models.BlobContainerEncryptionScope;
 import com.azure.storage.blob.models.CpkInfo;
+import com.azure.storage.blob.models.CpkScopeInfo;
 import com.azure.storage.blob.models.CustomerProvidedKey;
+import com.azure.storage.blob.models.EncryptionScope;
 import com.azure.storage.common.StorageSharedKeyCredential;
 import com.azure.storage.common.implementation.connectionstring.StorageAuthenticationSettings;
 import com.azure.storage.common.implementation.connectionstring.StorageConnectionString;
@@ -51,6 +54,8 @@ public final class BlobServiceClientBuilder {
     private String accountName;
 
     private CpkInfo customerProvidedKey;
+    private EncryptionScope encryptionScope;
+    private BlobContainerEncryptionScope blobContainerEncryptionScope;
     private StorageSharedKeyCredential storageSharedKeyCredential;
     private TokenCredential tokenCredential;
     private SasTokenCredential sasTokenCredential;
@@ -92,12 +97,18 @@ public final class BlobServiceClientBuilder {
                 + "anonymously. Please provide a form of authentication"));
         }
 
+        if (Objects.nonNull(customerProvidedKey) && Objects.nonNull(encryptionScope)) {
+            throw logger.logExceptionAsError(new IllegalArgumentException("Customer provided key and encryption"
+                + "scope cannot both be set"));
+        }
+
         BlobServiceVersion serviceVersion = version != null ? version : BlobServiceVersion.getLatest();
         HttpPipeline pipeline = (httpPipeline != null) ? httpPipeline : BuilderHelper.buildPipeline(
             storageSharedKeyCredential, tokenCredential, sasTokenCredential, endpoint, retryOptions, logOptions,
             httpClient, additionalPolicies, configuration, logger);
 
-        return new BlobServiceAsyncClient(pipeline, endpoint, serviceVersion, accountName, customerProvidedKey);
+        return new BlobServiceAsyncClient(pipeline, endpoint, serviceVersion, accountName, customerProvidedKey,
+            encryptionScope, blobContainerEncryptionScope);
     }
 
     /**
@@ -142,6 +153,30 @@ public final class BlobServiceClientBuilder {
                 .setEncryptionAlgorithm(customerProvidedKey.getEncryptionAlgorithm());
         }
 
+        return this;
+    }
+
+    /**
+     * Sets the {@link EncryptionScope encryption scope} that is used to encrypt blob contents on the server.
+     *
+     * @param encryptionScope Encryption scope containing the encryption key information.
+     * @return the updated BlobServiceClientBuilder object
+     */
+    public BlobServiceClientBuilder encryptionScope(EncryptionScope encryptionScope) {
+        this.encryptionScope = encryptionScope;
+        return this;
+    }
+
+    /**
+     * Sets the {@link BlobContainerEncryptionScope encryption scope} that is used to determine how blob contents are
+     * encrypted on the server.
+     *
+     * @param blobContainerEncryptionScope Encryption scope containing the encryption key information.
+     * @return the updated BlobServiceClientBuilder object
+     */
+    public BlobServiceClientBuilder blobContainerEncryptionScope(
+        BlobContainerEncryptionScope blobContainerEncryptionScope) {
+        this.blobContainerEncryptionScope = blobContainerEncryptionScope;
         return this;
     }