+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/models/BlockBlobItem.java
@@ -18,6 +18,7 @@ public class BlockBlobItem {
     private final byte[] contentMd5;
     private final Boolean isServerEncrypted;
     private final String encryptionKeySha256;
+    private final String encryptionScope;
 
     /**
      * Constructs a {@link BlockBlobItem}.
@@ -25,9 +26,10 @@ public class BlockBlobItem {
      * @param eTag ETag of the block blob.
      * @param lastModified Last modified time of the block blob.
      * @param contentMd5 Content MD5 of the block blob.
-     * @param isServerEncrypted Flag indicating if the page blob is encrypted on the server.
-     * @param encryptionKeySha256 The encryption key used to encrypt the page blob.
+     * @param isServerEncrypted Flag indicating if the block blob is encrypted on the server.
+     * @param encryptionKeySha256 The encryption key used to encrypt the block blob.
      */
+    @Deprecated
     public BlockBlobItem(final String eTag, final OffsetDateTime lastModified, final byte[] contentMd5,
         final boolean isServerEncrypted, final String encryptionKeySha256) {
         this.eTag = eTag;
@@ -35,6 +37,27 @@ public class BlockBlobItem {
         this.contentMd5 = CoreUtils.clone(contentMd5);
         this.isServerEncrypted = isServerEncrypted;
         this.encryptionKeySha256 = encryptionKeySha256;
+        this.encryptionScope = null;
+    }
+
+    /**
+     * Constructs a {@link BlockBlobItem}.
+     *
+     * @param eTag ETag of the block blob.
+     * @param lastModified Last modified time of the block blob.
+     * @param contentMd5 Content MD5 of the block blob.
+     * @param isServerEncrypted Flag indicating if the block blob is encrypted on the server.
+     * @param encryptionKeySha256 The encryption key used to encrypt the block blob.
+     * @param encryptionScope The encryption scope used to encrypt the block blob.
+     */
+    public BlockBlobItem(final String eTag, final OffsetDateTime lastModified, final byte[] contentMd5,
+        final boolean isServerEncrypted, final String encryptionKeySha256, final String encryptionScope) {
+        this.eTag = eTag;
+        this.lastModified = lastModified;
+        this.contentMd5 = CoreUtils.clone(contentMd5);
+        this.isServerEncrypted = isServerEncrypted;
+        this.encryptionKeySha256 = encryptionKeySha256;
+        this.encryptionScope = encryptionScope;
     }
 
     /**
@@ -63,6 +86,13 @@ public class BlockBlobItem {
      */
     public String getEncryptionKeySha256() {
         return encryptionKeySha256;
+    }
+
+    /**
+     * @return the encryption scope used to encrypt the block blob
+     */
+    public String getEncryptionScope() {
+        return encryptionScope;
     }
 
     /**