+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/models/PageBlobItem.java
@@ -18,6 +18,7 @@ public class PageBlobItem {
     private final byte[] contentMd5;
     private final Boolean isServerEncrypted;
     private final String encryptionKeySha256;
+    private final String encryptionScope;
     private final Long blobSequenceNumber;
 
     /**
@@ -30,6 +31,7 @@ public class PageBlobItem {
      * @param encryptionKeySha256 The encryption key used to encrypt the page blob.
      * @param blobSequenceNumber The current sequence number for the page blob.
      */
+    @Deprecated
     public PageBlobItem(final String eTag, final OffsetDateTime lastModified, final byte[] contentMd5,
         final Boolean isServerEncrypted, final String encryptionKeySha256, final Long blobSequenceNumber) {
         this.eTag = eTag;
@@ -37,6 +39,30 @@ public class PageBlobItem {
         this.contentMd5 = CoreUtils.clone(contentMd5);
         this.isServerEncrypted = isServerEncrypted;
         this.encryptionKeySha256 = encryptionKeySha256;
+        this.encryptionScope = null;
+        this.blobSequenceNumber = blobSequenceNumber;
+    }
+
+    /**
+     * Constructs a {@link PageBlobItem}.
+     *
+     * @param eTag ETag of the page blob.
+     * @param lastModified Last modified time of the page blob.
+     * @param contentMd5 Content MD5 of the page blob.
+     * @param isServerEncrypted Flag indicating if the page blob is encrypted on the server.
+     * @param encryptionKeySha256 The encryption key used to encrypt the page blob.
+     * @param encryptionScope The encryption scope used to encrypt the page blob.
+     * @param blobSequenceNumber The current sequence number for the page blob.
+     */
+    public PageBlobItem(final String eTag, final OffsetDateTime lastModified, final byte[] contentMd5,
+        final Boolean isServerEncrypted, final String encryptionKeySha256, final String encryptionScope,
+        final Long blobSequenceNumber) {
+        this.eTag = eTag;
+        this.lastModified = lastModified;
+        this.contentMd5 = CoreUtils.clone(contentMd5);
+        this.isServerEncrypted = isServerEncrypted;
+        this.encryptionKeySha256 = encryptionKeySha256;
+        this.encryptionScope = encryptionScope;
         this.blobSequenceNumber = blobSequenceNumber;
     }
 
@@ -66,6 +92,13 @@ public class PageBlobItem {
      */
     public String getEncryptionKeySha256() {
         return encryptionKeySha256;
+    }
+
+    /**
+     * @return the encryption scope used to encrypt the page blob
+     */
+    public String getEncryptionScope() {
+        return encryptionScope;
     }
 
     /**